Reorganizing build and testing.
