+++ gulpfile.js
@@ -1,32 +1,47 @@
 var gulp = require('gulp');
 var ts = require('gulp-typescript');
 var mocha = require('gulp-mocha');
+var clean = require('gulp-clean');
+ 
+gulp.task('default', ['trasnpile-source', 'create-source-definition', 'trasnpile-test', 'run-test', 'clean-test', 'distribute-source']);
 
-gulp.task('default', ['trasnpileSource', 'trasnpileTest', 'runTest', 'distributeSource']);
-
-gulp.task('trasnpileSource', function () {
+gulp.task('trasnpile-source', function () {
     return gulp.src('Source/**/*')
                 .pipe(ts({
                     declaration: true
                 }))
                 .js
-                .pipe(gulp.dest('Transpilation/Source'));
+                .pipe(gulp.dest('distribution'));
 });
 
-gulp.task('trasnpileTest', ['trasnpileSource'], function () {
+gulp.task('create-source-definition', ['trasnpile-source'], function () {
+    return gulp.src('Source/Index.ts')
+                .pipe(ts({
+                    declaration: true
+                }))
+                .dts
+                .pipe(gulp.dest('distribution'));
+});
+
+gulp.task('trasnpile-test', ['create-source-definition'], function () {
     return gulp.src('Test/**/*')
                 .pipe(ts({
                     declaration: true
                 }))
                 .js
-                .pipe(gulp.dest('Transpilation/Test'));
+                .pipe(gulp.dest('test_transpiled'));
 });
 
-gulp.task('runTest', ['trasnpileTest'], function () {
-    return gulp.src('Transpilation/Test/**/*')
+gulp.task('run-test', ['trasnpile-test'], function () {
+    return gulp.src('test_transpiled/TestStarter.js')
                .pipe(mocha());
 });
 
-gulp.task('distributeSource', ['runTest'], function () {
+gulp.task('clean-test', ['run-test'], function () {
+  return gulp.src(['test_transpiled'], {read: false})
+    .pipe(clean());
+});
+
+gulp.task('distribute-source', ['clean-test'], function () {
 
 });
