+++ source/Index.ts
@@ -1,0 +1,9 @@
+//Exporting the kernel
+
+import { Kernel } from './kernel'
+import { ServicingStrategy } from './ServicingStrategy'
+
+
+export { Kernel as Kernel };
+export { ServicingStrategy as ServicingStrategy };
+
