+++ test/fake_types/FakeTypeA.ts
