+++ test/fake_types/FakeTypeB.ts
