+++ test/fake_types/FakeTypeC.ts
